backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.8.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
supabase>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.2
openai>=1.0.0
stripe>=7.0.0
